Add HomeController.
